src/requirements.txt
Django==1.11.3
django-crispy-forms==1.6.1
djangorestframework==3.7.3
olefile==0.44
Pillow==4.2.1
pytz==2017.2
